feat: move concat agents and time to jax utils
